Add music notes to save editor, hook GV Lobby's framerate cap, and organize the monolithic Engine.cpp (#482)
* Add music notes to the save editor

* Release GV framerate cap except for in static scenes

* Split up and organize the monolithic Engine.cpp
